+++ mne_connectivity/__init__.py
@@ -15,6 +15,6 @@ from .base import (Connectivity, EpochConnectivity, EpochSpectralConnectivity,
 from .effective import phase_slope_index
 from .envelope import envelope_correlation, symmetric_orth
 from .io import read_connectivity
-from .spectral import spectral_connectivity, spectral_connectivity_epochs
+from .spectral import spectral_connectivity_time, spectral_connectivity_epochs
 from .vector_ar import vector_auto_regression, select_order
 from .utils import check_indices, degree, seed_target_indices
