+++ mne_connectivity/spectral/__init__.py
@@ -1,0 +1,2 @@
+from .epochs import spectral_connectivity_epochs
+from .time import spectral_connectivity_time
