+++ mne_connectivity/spectral/epochs.py
@@ -1,5 +1,6 @@
 # Authors: Martin Luessi <mluessi@nmr.mgh.harvard.edu>
 #          Denis A. Engemann <denis.engemann@gmail.com>
+#          Adam Li <adam2392@gmail.com>
 #
 # License: BSD (3-clause)
 
@@ -7,22 +8,182 @@ from functools import partial
 from inspect import getmembers
 
 import numpy as np
-import xarray as xr
 from mne.epochs import BaseEpochs
 from mne.fixes import _get_args
 from mne.parallel import parallel_func
 from mne.source_estimate import _BaseSourceEstimate
-from mne.time_frequency.multitaper import (_compute_mt_params, _csd_from_mt,
+from mne.time_frequency.multitaper import (_csd_from_mt,
                                            _mt_spectra, _psd_from_mt,
                                            _psd_from_mt_adaptive)
 from mne.time_frequency.tfr import cwt, morlet
-from mne.time_frequency import (tfr_array_morlet, tfr_array_multitaper)
-from mne.utils import (_arange_div, _check_option, _time_mask, logger, warn)
+from mne.time_frequency.multitaper import _compute_mt_params
+from mne.utils import (_arange_div, _check_option, logger, warn, _time_mask)
 
-from .base import (EpochSpectroTemporalConnectivity, SpectralConnectivity,
-                   SpectroTemporalConnectivity)
-from .smooth import _create_kernel, _smooth_spectra
-from .utils import check_indices, fill_doc
+from ..base import (SpectralConnectivity, SpectroTemporalConnectivity)
+from ..utils import fill_doc, check_indices
+
+
+def _prepare_connectivity(epoch_block, times_in, tmin, tmax,
+                          fmin, fmax, sfreq, indices,
+                          mode, fskip, n_bands,
+                          cwt_freqs, faverage):
+    """Check and precompute dimensions of results data."""
+    from scipy.fft import rfftfreq
+    first_epoch = epoch_block[0]
+
+    # get the data size and time scale
+    n_signals, n_times_in, times_in, warn_times = _get_and_verify_data_sizes(
+        first_epoch, sfreq, times=times_in)
+
+    n_times_in = len(times_in)
+
+    if tmin is not None and tmin < times_in[0]:
+        warn('start time tmin=%0.2f s outside of the time scope of the data '
+             '[%0.2f s, %0.2f s]' % (tmin, times_in[0], times_in[-1]))
+    if tmax is not None and tmax > times_in[-1]:
+        warn('stop time tmax=%0.2f s outside of the time scope of the data '
+             '[%0.2f s, %0.2f s]' % (tmax, times_in[0], times_in[-1]))
+
+    mask = _time_mask(times_in, tmin, tmax, sfreq=sfreq)
+    tmin_idx, tmax_idx = np.where(mask)[0][[0, -1]]
+    tmax_idx += 1
+    tmin_true = times_in[tmin_idx]
+    tmax_true = times_in[tmax_idx - 1]  # time of last point used
+
+    times = times_in[tmin_idx:tmax_idx]
+    n_times = len(times)
+
+    if indices is None:
+        logger.info('only using indices for lower-triangular matrix')
+        # only compute r for lower-triangular region
+        indices_use = np.tril_indices(n_signals, -1)
+    else:
+        indices_use = check_indices(indices)
+
+    # number of connectivities to compute
+    n_cons = len(indices_use[0])
+
+    logger.info('    computing connectivity for %d connections'
+                % n_cons)
+    logger.info('    using t=%0.3fs..%0.3fs for estimation (%d points)'
+                % (tmin_true, tmax_true, n_times))
+
+    # get frequencies of interest for the different modes
+    if mode in ('multitaper', 'fourier'):
+        # fmin fmax etc is only supported for these modes
+        # decide which frequencies to keep
+        freqs_all = rfftfreq(n_times, 1. / sfreq)
+    elif mode == 'cwt_morlet':
+        # cwt_morlet mode
+        if cwt_freqs is None:
+            raise ValueError('define frequencies of interest using '
+                             'cwt_freqs')
+        else:
+            cwt_freqs = cwt_freqs.astype(np.float64)
+        if any(cwt_freqs > (sfreq / 2.)):
+            raise ValueError('entries in cwt_freqs cannot be '
+                             'larger than Nyquist (sfreq / 2)')
+        freqs_all = cwt_freqs
+    else:
+        raise ValueError('mode has an invalid value')
+
+    # check that fmin corresponds to at least 5 cycles
+    dur = float(n_times) / sfreq
+    five_cycle_freq = 5. / dur
+    if len(fmin) == 1 and fmin[0] == -np.inf:
+        # we use the 5 cycle freq. as default
+        fmin = np.array([five_cycle_freq])
+    else:
+        if np.any(fmin < five_cycle_freq):
+            warn('fmin=%0.3f Hz corresponds to %0.3f < 5 cycles '
+                 'based on the epoch length %0.3f sec, need at least %0.3f '
+                 'sec epochs or fmin=%0.3f. Spectrum estimate will be '
+                 'unreliable.' % (np.min(fmin), dur * np.min(fmin), dur,
+                                  5. / np.min(fmin), five_cycle_freq))
+
+    # create a frequency mask for all bands
+    freq_mask = np.zeros(len(freqs_all), dtype=bool)
+    for f_lower, f_upper in zip(fmin, fmax):
+        freq_mask |= ((freqs_all >= f_lower) & (freqs_all <= f_upper))
+
+    # possibly skip frequency points
+    for pos in range(fskip):
+        freq_mask[pos + 1::fskip + 1] = False
+
+    # the frequency points where we compute connectivity
+    freqs = freqs_all[freq_mask]
+    n_freqs = len(freqs)
+
+    # get the freq. indices and points for each band
+    freq_idx_bands = [np.where((freqs >= fl) & (freqs <= fu))[0]
+                      for fl, fu in zip(fmin, fmax)]
+    freqs_bands = [freqs[freq_idx] for freq_idx in freq_idx_bands]
+
+    # make sure we don't have empty bands
+    for i, n_f_band in enumerate([len(f) for f in freqs_bands]):
+        if n_f_band == 0:
+            raise ValueError('There are no frequency points between '
+                             '%0.1fHz and %0.1fHz. Change the band '
+                             'specification (fmin, fmax) or the '
+                             'frequency resolution.'
+                             % (fmin[i], fmax[i]))
+    if n_bands == 1:
+        logger.info('    frequencies: %0.1fHz..%0.1fHz (%d points)'
+                    % (freqs_bands[0][0], freqs_bands[0][-1],
+                       n_freqs))
+    else:
+        logger.info('    computing connectivity for the bands:')
+        for i, bfreqs in enumerate(freqs_bands):
+            logger.info('     band %d: %0.1fHz..%0.1fHz '
+                        '(%d points)' % (i + 1, bfreqs[0],
+                                         bfreqs[-1], len(bfreqs)))
+    if faverage:
+        logger.info('    connectivity scores will be averaged for '
+                    'each band')
+
+    return (n_cons, times, n_times, times_in, n_times_in, tmin_idx,
+            tmax_idx, n_freqs, freq_mask, freqs, freqs_bands, freq_idx_bands,
+            n_signals, indices_use, warn_times)
+
+
+def _assemble_spectral_params(mode, n_times, mt_adaptive, mt_bandwidth, sfreq,
+                              mt_low_bias, cwt_n_cycles, cwt_freqs,
+                              freqs, freq_mask):
+    """Prepare time-frequency decomposition."""
+    spectral_params = dict(
+        eigvals=None, window_fun=None, wavelets=None)
+    n_tapers = None
+    n_times_spectrum = 0
+    if mode == 'multitaper':
+        window_fun, eigvals, mt_adaptive = _compute_mt_params(
+            n_times, sfreq, mt_bandwidth, mt_low_bias, mt_adaptive)
+        spectral_params.update(window_fun=window_fun, eigvals=eigvals)
+    elif mode == 'fourier':
+        logger.info('    using FFT with a Hanning window to estimate '
+                    'spectra')
+        spectral_params.update(window_fun=np.hanning(n_times), eigvals=1.)
+    elif mode == 'cwt_morlet':
+        logger.info('    using CWT with Morlet wavelets to estimate '
+                    'spectra')
+
+        # reformat cwt_n_cycles if we have removed some frequencies
+        # using fmin, fmax, fskip
+        cwt_n_cycles = np.array((cwt_n_cycles,), dtype=float).ravel()
+        if len(cwt_n_cycles) > 1:
+            if len(cwt_n_cycles) != len(cwt_freqs):
+                raise ValueError('cwt_n_cycles must be float or an '
+                                 'array with the same size as cwt_freqs')
+            cwt_n_cycles = cwt_n_cycles[freq_mask]
+
+        # get the Morlet wavelets
+        spectral_params.update(
+            wavelets=morlet(sfreq, freqs,
+                            n_cycles=cwt_n_cycles, zero_mean=True))
+        n_times_spectrum = n_times
+    else:
+        raise ValueError('mode has an invalid value')
+    return spectral_params, mt_adaptive, n_times_spectrum, n_tapers
+
 
 ########################################################################
 # Various connectivity estimators
@@ -561,14 +722,14 @@ def _check_estimators(method, mode):
 
 
 @fill_doc
-def spectral_connectivity(data, names=None, method='coh', indices=None,
-                          sfreq=2 * np.pi,
-                          mode='multitaper', fmin=None, fmax=np.inf,
-                          fskip=0, faverage=False, tmin=None, tmax=None,
-                          mt_bandwidth=None, mt_adaptive=False,
-                          mt_low_bias=True, cwt_freqs=None,
-                          cwt_n_cycles=7, block_size=1000, n_jobs=1,
-                          verbose=None):
+def spectral_connectivity_epochs(data, names=None, method='coh', indices=None,
+                                 sfreq=2 * np.pi,
+                                 mode='multitaper', fmin=None, fmax=np.inf,
+                                 fskip=0, faverage=False, tmin=None, tmax=None,
+                                 mt_bandwidth=None, mt_adaptive=False,
+                                 mt_low_bias=True, cwt_freqs=None,
+                                 cwt_n_cycles=7, block_size=1000, n_jobs=1,
+                                 verbose=None):
     """Compute frequency- and time-frequency-domain connectivity measures.
 
     The connectivity method(s) are specified using the "method" parameter.
@@ -994,529 +1155,3 @@ def spectral_connectivity(data, names=None, method='coh', indices=None,
         conn_list = conn_list[0]
 
     return conn_list
-
-
-@fill_doc
-def spectral_connectivity_epochs(data, names=None, method='coh', indices=None,
-                                 sfreq=2 * np.pi, foi=None, sm_times=.5,
-                                 sm_freqs=1, sm_kernel='hanning',
-                                 mode='cwt_morlet', mt_bandwidth=None,
-                                 freqs=None, n_cycles=7, decim=1,
-                                 block_size=None, n_jobs=1,
-                                 verbose=None):
-    """Compute frequency- and time-frequency-domain connectivity measures.
-
-    This method computes single-Epoch time-resolved spectral connectivity.
-
-    The connectivity method(s) are specified using the "method" parameter.
-    All methods are based on estimates of the cross- and power spectral
-    densities (CSD/PSD) Sxy and Sxx, Syy.
-
-    Parameters
-    ----------
-    data : Epochs
-        The data from which to compute connectivity.
-    %(names)s
-    method : str | list of str
-        Connectivity measure(s) to compute. These can be ``['coh', 'plv',
-        'sxy']``. These are:
-
-            * 'coh' : Coherence
-            * 'plv' : Phase-Locking Value (PLV)
-            * 'sxy' : Cross-spectrum
-
-        By default, the coherence is used.
-    indices : tuple of array | None
-        Two arrays with indices of connections for which to compute
-        connectivity. I.e. it is a ``(n_pairs, 2)`` array essentially.
-        If None, all connections are computed.
-    sfreq : float
-        The sampling frequency.
-    foi : array_like | None
-        Extract frequencies of interest. This parameters should be an array of
-        shapes (n_foi, 2) defining where each band of interest start and
-        finish.
-    sm_times : float | .5
-        Number of points to consider for the temporal smoothing in seconds. By
-        default, a 500ms smoothing is used.
-    sm_freqs : int | 1
-        Number of points for frequency smoothing. By default, 1 is used which
-        is equivalent to no smoothing
-    kernel : {'square', 'hanning'}
-        Kernel type to use. Choose either 'square' or 'hanning'
-    mode : str, optional
-        Spectrum estimation mode can be either: 'multitaper', or
-        'cwt_morlet'.
-    mt_bandwidth : float | None
-        The bandwidth of the multitaper windowing function in Hz.
-        Only used in 'multitaper' mode.
-    freqs : array
-        Array of frequencies of interest for use in time-frequency
-        decomposition method (specified by ``mode``).
-    n_cycles : float | array of float
-        Number of cycles for use in time-frequency decomposition method
-        (specified by ``mode``). Fixed number or one per frequency.
-    decim : int | 1
-        To reduce memory usage, decimation factor after time-frequency
-        decomposition. default 1 If int, returns tfr[…, ::decim]. If slice,
-        returns tfr[…, decim].
-    block_size : int
-        How many connections to compute at once (higher numbers are faster
-        but require more memory).
-    n_jobs : int
-        How many epochs to process in parallel.
-    %(verbose)s
-
-    Returns
-    -------
-    con : array | instance of Connectivity
-        Computed connectivity measure(s). Either an instance of
-        ``SpectralConnectivity`` or ``SpectroTemporalConnectivity``.
-        The shape of each connectivity dataset is either
-        (n_signals ** 2, n_freqs) mode: 'multitaper' or 'fourier'
-        (n_signals ** 2, n_freqs, n_times) mode: 'cwt_morlet'
-        when "indices" is None, or
-        (n_con, n_freqs) mode: 'multitaper' or 'fourier'
-        (n_con, n_freqs, n_times) mode: 'cwt_morlet'
-        when "indices" is specified and "n_con = len(indices[0])".
-
-    See Also
-    --------
-    mne_connectivity.spectral_connectivity
-    mne_connectivity.SpectralConnectivity
-    mne_connectivity.SpectroTemporalConnectivity
-
-    Notes
-    -----
-    This function was originally implemented in ``frites``.
-    x_s and x_t -> source_idx, and target_idx
-    sf -> sfreq
-
-    """
-    events = None
-    event_id = None
-    # extract data from Epochs object
-    names = data.ch_names
-    times = data.times  # input times for Epochs input type
-    sfreq = data.info['sfreq']
-    events = data.events
-    event_id = data.event_id
-    n_epochs, n_signals, n_times = data.get_data().shape
-    # Extract metadata from the Epochs data structure.
-    # Make Annotations persist through by adding them to the metadata.
-    if hasattr(data, 'annotations'):
-        data.add_annotations_to_metadata()
-    metadata = data.metadata
-    data = data.get_data()
-
-    # convert kernel width in time to samples
-    if isinstance(sm_times, (int, float)):
-        sm_times = int(np.round(sm_times * sfreq))
-
-    # convert frequency smoothing from hz to samples
-    if isinstance(sm_freqs, (int, float)):
-        sm_freqs = int(np.round(max(sm_freqs, 1)))
-
-    # temporal decimation
-    if isinstance(decim, int):
-        times = times[::decim]
-        sm_times = int(np.round(sm_times / decim))
-        sm_times = max(sm_times, 1)
-
-    # Create smoothing kernel
-    kernel = _create_kernel(sm_times, sm_freqs, kernel=sm_kernel)
-
-    # get indices of pairs of (group) regions
-    roi = names  # ch_names
-    if indices is None:
-        # roi_gp and roi_idx
-        roi_gp, _ = roi, np.arange(len(roi)).reshape(-1, 1)
-
-        # get pairs for directed / undirected conn
-        source_idx, target_idx = np.triu_indices(len(roi_gp), k=0)
-    else:
-        indices_use = check_indices(indices)
-        source_idx = [x[0] for x in indices_use]
-        target_idx = [x[1] for x in indices_use]
-        roi_gp, _ = roi, np.arange(len(roi)).reshape(-1, 1)
-    n_pairs = len(source_idx)
-
-    # frequency checking
-    if freqs is not None:
-        # check for single frequency
-        if isinstance(freqs, (int, float)):
-            freqs = [freqs]
-        # array conversion
-        freqs = np.asarray(freqs)
-        # check order for multiple frequencies
-        if len(freqs) >= 2:
-            delta_f = np.diff(freqs)
-            increase = np.all(delta_f > 0)
-            assert increase, "Frequencies should be in increasing order"
-
-        # frequency mean
-        if foi is None:
-            foi_idx = foi_s = foi_e = None
-            f_vec = freqs
-        else:
-            _f = xr.DataArray(np.arange(len(freqs)), dims=('freqs',),
-                              coords=(freqs,))
-            foi_s = _f.sel(freqs=foi[:, 0], method='nearest').data
-            foi_e = _f.sel(freqs=foi[:, 1], method='nearest').data
-            foi_idx = np.c_[foi_s, foi_e]
-            f_vec = freqs[foi_idx].mean(1)
-
-    # build block size indices
-    if isinstance(block_size, int) and (block_size > 1):
-        blocks = np.array_split(np.arange(n_epochs), block_size)
-    else:
-        blocks = [np.arange(n_epochs)]
-
-    n_freqs = len(f_vec)
-
-    # compute coherence on blocks of trials
-    conn = np.zeros((n_epochs, n_pairs, n_freqs, len(times)))
-    logger.info('Connectivity computation...')
-
-    # parameters to pass to the connectivity function
-    call_params = dict(
-        method=method, kernel=kernel, foi_idx=foi_idx,
-        source_idx=source_idx, target_idx=target_idx,
-        mode=mode, sfreq=sfreq, freqs=freqs, n_cycles=n_cycles,
-        mt_bandwidth=mt_bandwidth,
-        decim=decim, kw_cwt={}, kw_mt={}, n_jobs=n_jobs,
-        verbose=verbose)
-
-    for epoch_idx in blocks:
-        # compute time-resolved spectral connectivity
-        conn_tr = _spectral_connectivity(data[epoch_idx, ...], **call_params)
-
-        # merge results
-        conn[epoch_idx, ...] = np.stack(conn_tr, axis=1)
-
-    # create a Connectivity container
-    indices = 'symmetric'
-    conn = EpochSpectroTemporalConnectivity(
-        conn, freqs=freqs, times=times,
-        n_nodes=n_signals, names=names, indices=indices, method=method,
-        spec_method=mode, events=events, event_id=event_id, metadata=metadata)
-
-    return conn
-
-
-def _spectral_connectivity(data, method, kernel, foi_idx,
-                           source_idx, target_idx,
-                           mode, sfreq, freqs, n_cycles, mt_bandwidth=None,
-                           decim=1, kw_cwt={}, kw_mt={}, n_jobs=1,
-                           verbose=False):
-    """EStimate time-resolved connectivity for one epoch.
-
-    See spectral_connectivity_epoch."""
-    n_pairs = len(source_idx)
-
-    # first compute time-frequency decomposition
-    if mode == 'cwt_morlet':
-        out = tfr_array_morlet(
-            data, sfreq, freqs, n_cycles=n_cycles, output='complex',
-            decim=decim, n_jobs=n_jobs, **kw_cwt)
-    elif mode == 'multitaper':
-        # In case multiple values are provided for mt_bandwidth
-        # the MT decomposition is done separatedly for each
-        # Frequency center
-        if isinstance(mt_bandwidth, (list, tuple, np.ndarray)):
-            # Arrays freqs, n_cycles, mt_bandwidth should have the same size
-            assert len(freqs) == len(n_cycles) == len(mt_bandwidth)
-            out = []
-            for f_c, n_c, mt in zip(freqs, n_cycles, mt_bandwidth):
-                out += [tfr_array_multitaper(
-                    data, sfreq, [f_c], n_cycles=float(n_c), time_bandwidth=mt,
-                    output='complex', decim=decim, n_jobs=n_jobs, **kw_mt)]
-            out = np.stack(out, axis=2).squeeze()
-        elif isinstance(mt_bandwidth, (type(None), int, float)):
-            out = tfr_array_multitaper(
-                data, sfreq, freqs, n_cycles=n_cycles,
-                time_bandwidth=mt_bandwidth, output='complex', decim=decim,
-                n_jobs=n_jobs, **kw_mt)
-
-    # get the supported connectivity function
-    conn_func = {'coh': _coh, 'plv': _plv, 'sxy': _cs}[method]
-
-    # computes conn across trials
-    this_conn = conn_func(out, kernel, foi_idx, source_idx, target_idx,
-                          n_jobs=n_jobs, verbose=verbose, total=n_pairs)
-    return this_conn
-
-
-def _prepare_connectivity(epoch_block, times_in, tmin, tmax,
-                          fmin, fmax, sfreq, indices,
-                          mode, fskip, n_bands,
-                          cwt_freqs, faverage):
-    """Check and precompute dimensions of results data."""
-    from scipy.fft import rfftfreq
-    first_epoch = epoch_block[0]
-
-    # get the data size and time scale
-    n_signals, n_times_in, times_in, warn_times = _get_and_verify_data_sizes(
-        first_epoch, sfreq, times=times_in)
-
-    n_times_in = len(times_in)
-
-    if tmin is not None and tmin < times_in[0]:
-        warn('start time tmin=%0.2f s outside of the time scope of the data '
-             '[%0.2f s, %0.2f s]' % (tmin, times_in[0], times_in[-1]))
-    if tmax is not None and tmax > times_in[-1]:
-        warn('stop time tmax=%0.2f s outside of the time scope of the data '
-             '[%0.2f s, %0.2f s]' % (tmax, times_in[0], times_in[-1]))
-
-    mask = _time_mask(times_in, tmin, tmax, sfreq=sfreq)
-    tmin_idx, tmax_idx = np.where(mask)[0][[0, -1]]
-    tmax_idx += 1
-    tmin_true = times_in[tmin_idx]
-    tmax_true = times_in[tmax_idx - 1]  # time of last point used
-
-    times = times_in[tmin_idx:tmax_idx]
-    n_times = len(times)
-
-    if indices is None:
-        logger.info('only using indices for lower-triangular matrix')
-        # only compute r for lower-triangular region
-        indices_use = np.tril_indices(n_signals, -1)
-    else:
-        indices_use = check_indices(indices)
-
-    # number of connectivities to compute
-    n_cons = len(indices_use[0])
-
-    logger.info('    computing connectivity for %d connections'
-                % n_cons)
-    logger.info('    using t=%0.3fs..%0.3fs for estimation (%d points)'
-                % (tmin_true, tmax_true, n_times))
-
-    # get frequencies of interest for the different modes
-    if mode in ('multitaper', 'fourier'):
-        # fmin fmax etc is only supported for these modes
-        # decide which frequencies to keep
-        freqs_all = rfftfreq(n_times, 1. / sfreq)
-    elif mode == 'cwt_morlet':
-        # cwt_morlet mode
-        if cwt_freqs is None:
-            raise ValueError('define frequencies of interest using '
-                             'cwt_freqs')
-        else:
-            cwt_freqs = cwt_freqs.astype(np.float64)
-        if any(cwt_freqs > (sfreq / 2.)):
-            raise ValueError('entries in cwt_freqs cannot be '
-                             'larger than Nyquist (sfreq / 2)')
-        freqs_all = cwt_freqs
-    else:
-        raise ValueError('mode has an invalid value')
-
-    # check that fmin corresponds to at least 5 cycles
-    dur = float(n_times) / sfreq
-    five_cycle_freq = 5. / dur
-    if len(fmin) == 1 and fmin[0] == -np.inf:
-        # we use the 5 cycle freq. as default
-        fmin = np.array([five_cycle_freq])
-    else:
-        if np.any(fmin < five_cycle_freq):
-            warn('fmin=%0.3f Hz corresponds to %0.3f < 5 cycles '
-                 'based on the epoch length %0.3f sec, need at least %0.3f '
-                 'sec epochs or fmin=%0.3f. Spectrum estimate will be '
-                 'unreliable.' % (np.min(fmin), dur * np.min(fmin), dur,
-                                  5. / np.min(fmin), five_cycle_freq))
-
-    # create a frequency mask for all bands
-    freq_mask = np.zeros(len(freqs_all), dtype=bool)
-    for f_lower, f_upper in zip(fmin, fmax):
-        freq_mask |= ((freqs_all >= f_lower) & (freqs_all <= f_upper))
-
-    # possibly skip frequency points
-    for pos in range(fskip):
-        freq_mask[pos + 1::fskip + 1] = False
-
-    # the frequency points where we compute connectivity
-    freqs = freqs_all[freq_mask]
-    n_freqs = len(freqs)
-
-    # get the freq. indices and points for each band
-    freq_idx_bands = [np.where((freqs >= fl) & (freqs <= fu))[0]
-                      for fl, fu in zip(fmin, fmax)]
-    freqs_bands = [freqs[freq_idx] for freq_idx in freq_idx_bands]
-
-    # make sure we don't have empty bands
-    for i, n_f_band in enumerate([len(f) for f in freqs_bands]):
-        if n_f_band == 0:
-            raise ValueError('There are no frequency points between '
-                             '%0.1fHz and %0.1fHz. Change the band '
-                             'specification (fmin, fmax) or the '
-                             'frequency resolution.'
-                             % (fmin[i], fmax[i]))
-    if n_bands == 1:
-        logger.info('    frequencies: %0.1fHz..%0.1fHz (%d points)'
-                    % (freqs_bands[0][0], freqs_bands[0][-1],
-                       n_freqs))
-    else:
-        logger.info('    computing connectivity for the bands:')
-        for i, bfreqs in enumerate(freqs_bands):
-            logger.info('     band %d: %0.1fHz..%0.1fHz '
-                        '(%d points)' % (i + 1, bfreqs[0],
-                                         bfreqs[-1], len(bfreqs)))
-    if faverage:
-        logger.info('    connectivity scores will be averaged for '
-                    'each band')
-
-    return (n_cons, times, n_times, times_in, n_times_in, tmin_idx,
-            tmax_idx, n_freqs, freq_mask, freqs, freqs_bands, freq_idx_bands,
-            n_signals, indices_use, warn_times)
-
-
-def _assemble_spectral_params(mode, n_times, mt_adaptive, mt_bandwidth, sfreq,
-                              mt_low_bias, cwt_n_cycles, cwt_freqs,
-                              freqs, freq_mask):
-    """Prepare time-frequency decomposition."""
-    spectral_params = dict(
-        eigvals=None, window_fun=None, wavelets=None)
-    n_tapers = None
-    n_times_spectrum = 0
-    if mode == 'multitaper':
-        window_fun, eigvals, mt_adaptive = _compute_mt_params(
-            n_times, sfreq, mt_bandwidth, mt_low_bias, mt_adaptive)
-        spectral_params.update(window_fun=window_fun, eigvals=eigvals)
-    elif mode == 'fourier':
-        logger.info('    using FFT with a Hanning window to estimate '
-                    'spectra')
-        spectral_params.update(window_fun=np.hanning(n_times), eigvals=1.)
-    elif mode == 'cwt_morlet':
-        logger.info('    using CWT with Morlet wavelets to estimate '
-                    'spectra')
-
-        # reformat cwt_n_cycles if we have removed some frequencies
-        # using fmin, fmax, fskip
-        cwt_n_cycles = np.array((cwt_n_cycles,), dtype=float).ravel()
-        if len(cwt_n_cycles) > 1:
-            if len(cwt_n_cycles) != len(cwt_freqs):
-                raise ValueError('cwt_n_cycles must be float or an '
-                                 'array with the same size as cwt_freqs')
-            cwt_n_cycles = cwt_n_cycles[freq_mask]
-
-        # get the Morlet wavelets
-        spectral_params.update(
-            wavelets=morlet(sfreq, freqs,
-                            n_cycles=cwt_n_cycles, zero_mean=True))
-        n_times_spectrum = n_times
-    else:
-        raise ValueError('mode has an invalid value')
-    return spectral_params, mt_adaptive, n_times_spectrum, n_tapers
-
-
-###############################################################################
-###############################################################################
-#                               TIME-RESOLVED CORE FUNCTIONS
-###############################################################################
-###############################################################################
-
-def _coh(w, kernel, foi_idx, source_idx, target_idx, n_jobs, verbose, total):
-    """Pairwise coherence."""
-    # auto spectra (faster that w * w.conj())
-    s_auto = w.real ** 2 + w.imag ** 2
-
-    # smooth the auto spectra
-    s_auto = _smooth_spectra(s_auto, kernel)
-
-    # define the pairwise coherence
-    def pairwise_coh(w_x, w_y):
-        # computes the coherence
-        s_xy = w[:, w_y, :, :] * np.conj(w[:, w_x, :, :])
-        s_xy = _smooth_spectra(s_xy, kernel)
-        s_xx = s_auto[:, w_x, :, :]
-        s_yy = s_auto[:, w_y, :, :]
-        out = np.abs(s_xy) ** 2 / (s_xx * s_yy)
-        # mean inside frequency sliding window (if needed)
-        if isinstance(foi_idx, np.ndarray):
-            return _foi_average(out, foi_idx)
-        else:
-            return out
-
-    # define the function to compute in parallel
-    parallel, p_fun, n_jobs = parallel_func(
-        pairwise_coh, n_jobs=n_jobs, verbose=verbose, total=total)
-
-    # compute the single trial coherence
-    return parallel(p_fun(s, t) for s, t in zip(source_idx, target_idx))
-
-
-def _plv(w, kernel, foi_idx, source_idx, target_idx, n_jobs, verbose, total):
-    """Pairwise phase-locking value."""
-    # define the pairwise plv
-    def pairwise_plv(w_x, w_y):
-        # computes the plv
-        s_xy = w[:, w_y, :, :] * np.conj(w[:, w_x, :, :])
-        # complex exponential of phase differences
-        exp_dphi = s_xy / np.abs(s_xy)
-        # smooth e^(-i*\delta\phi)
-        exp_dphi = _smooth_spectra(exp_dphi, kernel)
-        # computes plv
-        out = np.abs(exp_dphi)
-        # mean inside frequency sliding window (if needed)
-        if isinstance(foi_idx, np.ndarray):
-            return _foi_average(out, foi_idx)
-        else:
-            return out
-
-    # define the function to compute in parallel
-    parallel, p_fun, n_jobs = parallel_func(
-        pairwise_plv, n_jobs=n_jobs, verbose=verbose, total=total)
-
-    # compute the single trial coherence
-    return parallel(p_fun(s, t) for s, t in zip(source_idx, target_idx))
-
-
-def _cs(w, kernel, foi_idx, source_idx, target_idx, n_jobs, verbose, total):
-    """Pairwise cross-spectra."""
-    # define the pairwise cross-spectra
-    def pairwise_cs(w_x, w_y):
-        #  computes the cross-spectra
-        out = w[:, w_x, :, :] * np.conj(w[:, w_y, :, :])
-        out = _smooth_spectra(out, kernel)
-        if foi_idx is not None:
-            return _foi_average(out, foi_idx)
-        else:
-            return out
-
-    # define the function to compute in parallel
-    parallel, p_fun, n_jobs = parallel_func(
-        pairwise_cs, n_jobs=n_jobs, verbose=verbose, total=total)
-
-    # compute the single trial coherence
-    return parallel(p_fun(s, t) for s, t in zip(source_idx, target_idx))
-
-
-def _foi_average(conn, foi_idx):
-    """Average inside frequency bands.
-
-    The frequency dimension should be located at -2.
-
-    Parameters
-    ----------
-    conn : np.ndarray
-        Array of shape (..., n_freqs, n_times)
-    foi_idx : array_like
-        Array of indices describing frequency bounds of shape (n_foi, 2)
-
-    Returns
-    -------
-    conn_f : np.ndarray
-        Array of shape (..., n_foi, n_times)
-    """
-    # get the number of foi
-    n_foi = foi_idx.shape[0]
-
-    # get input shape and replace n_freqs with the number of foi
-    sh = list(conn.shape)
-    sh[-2] = n_foi
-
-    # compute average
-    conn_f = np.zeros(sh, dtype=conn.dtype)
-    for n_f, (f_s, f_e) in enumerate(foi_idx):
-        conn_f[..., n_f, :] = conn[..., f_s:f_e, :].mean(-2)
-    return conn_f
