+++ mne_connectivity/spectral/smooth.py
